app/requirements.txt
torch
torchvision
pillow
tqdm
argparse
transformers
pathlib
flask